// src/app/main/pages/laboratorio/model/ConceptoLiquidacionDto.ts
export class ConceptoLiquidacionDto{


    idConceptoLiquidacion: number;

    idIva: number;

    nombreConceptoLiq: string;

    descripcionConceptoLiq: string;

    precio: number;

    fechaConceptoLiq: Date;

    idUsuarioConceptoLiq: number;


    estadoConceptoLiq: string;

    estado: boolean;

    fechaTxt: string;




}